Modify Cloudflare Pages deployment workflow
This commit modifies the existing workflow file pages-deployment.yaml for deploying the website to Cloudflare Pages. The name field of the deployment job is updated to Direct Upload to Cloudflare Pages to better reflect the purpose of the job.

Additionally, the build step is uncommented and now runs npm install and npm run build to build the website before publishing it to Cloudflare Pages. This ensures that any changes made to the website are reflected in the deployed version.

Overall, these modifications improve the deployment workflow and ensure that the website is deployed with the latest changes.